+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-rc.1] - 2020-06-18
+### Added
+- New messages that change karma
+- New /help command
+- Add welcome messages
+### Changed
+- Bug fixes: all data is automaticly migrated after group migration to supergroup
+- Typo fixes
+- Now you can't use some bot commands in private messages
+
 ## [0.1.0-beta.3] - 2020-06-16
 ### Changed
 - Typo fixed
@@ -28,3 +38,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [0.1.0-beta.2]: https://github.com/sandsbit/skarmabot/compare/master...v0.1.0-beta.2
 [0.1.0-beta.3]: https://github.com/sandsbit/skarmabot/compare/master...v0.1.0-beta.3
+[0.1.0-rc.1]: https://github.com/sandsbit/skarmabot/compare/master...v0.1.0-rc.1